% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_arg_num}
\alias{check_arg_num}
\title{Check numeric arguments}
\usage{
check_arg_num(arg, arg_name)
}
\description{
This helper function checks whether an argument is a positive
number. Stops execution and prints the argument name to the user if an invalid
value is provided.
}
