% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_arg_numrange}
\alias{check_arg_numrange}
\title{Check numeric range arguments}
\usage{
check_arg_numrange(arg, arg_min, arg_max, arg_name)
}
\description{
This helper function checks whether an argument is a positive
number in the range [arg_min, arg_max]. Stops execution and prints the argument
name to the user if an invalid value is provided.
}
