% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_col_num}
\alias{check_col_num}
\title{#' Check numeric columns
#'
#' @description This helper function checks whether a column contains positive
#' numeric data. Stops execution and prints the column name to the user if a
#' column contains invalid data.
check_col_num <- function(col_values, col_name) {
  if (any(!is.numeric(col_values)) | any(col_values < 0)) {
    stop(paste("Error: column", col_name, "must contain positive numbers."))
  }
}}
\usage{
check_col_num(col_values, col_name)
}
\description{
This helper function checks whether a column contains positive
numeric data. Returns FALSE if data is valid, and TRUE if data is invalid.
}
\details{
#' Check numeric range columns
#'
#' @description This helper function checks whether a columns contains positive
#' numbers in the range [arg_min, arg_max]. Stops execution and prints the column
#' name to the user if it contains invalid values.
check_col_numrange <- function(col_values, col_min, col_max, col_name) {
  if (any(!is.numeric(col_values)) | any(col_values < col_min) | any(col_values > col_max)) {
    stop(paste("Error: column", col_name,"must contain numbers between", col_min, "and", col_max, "."))
  }
}
Check numeric columns
}
