% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_col_numrange}
\alias{check_col_numrange}
\title{Check numeric range columns}
\usage{
check_col_numrange(col_values, col_min, col_max, col_name)
}
\description{
This helper function checks whether a columns contains positive
numbers in the range [arg_min, arg_max]. Stops execution and prints the column
name to the user if it contains invalid values.
}
