% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input.R
\name{check_model_input}
\alias{check_model_input}
\title{Check model input}
\usage{
check_model_input(model_input)
}
\description{
This helper function wraps functions for checking whether column
values are valid and applies these to the model input dataframe
}
