% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_parameters.R
\name{draw_parameters}
\alias{draw_parameters}
\title{Draw parameters}
\usage{
draw_parameters(
  par_name,
  draw_method,
  par_vec = NULL,
  n_iterations,
  mean = NULL,
  sd = NULL,
  lower_bound = -Inf,
  upper_bound = Inf
)
}
\arguments{
\item{par_name}{A parameter name.}

\item{draw_method}{The method used to draw values.}

\item{par_vec}{A parameter vector.}

\item{n_iterations}{The number of values to draw.}

\item{mean}{Mean value for the `TruncNorm` draw method.}

\item{sd}{Standard deviation for the `TruncNorm` draw method.}

\item{lower_bound}{Upper bound for truncation for the `TruncNorm` draw method.}

\item{upper_bound}{Lower bound for truncation for the `TruncNorm` draw method.}
}
\description{
Draw parameters using three draw methods. Currently, the supported options are
`PointEstimate`, `Bootstrap` and `TruncNorm`. The `PointEstimate` option replicates
a single point estimate `n_iterations` times, while the `Bootstrap` option bootstraps
a parameter vector with replacement `n_iterations` times. The `TruncNorm` option samples
parameter values from a truncated normal distribution `n_iterations` using a `mean` and `sd`.
By default, the `TruncNorm` option assumes infinite lower and upper bounds, but these
can be specified by the user.
}
