% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfcm.R
\name{fcm}
\alias{fcm}
\title{Definition of the Flux Collision Model.}
\usage{
fcm(
  flux,
  a_macro,
  f_prop = 1,
  h_prop,
  h_prop_ref,
  rotor_d,
  rotor_d_ref,
  turb_dist,
  turb_dist_ref,
  turbs_e,
  turbs_e_ref,
  p_col,
  ...
)
}
\arguments{
\item{flux}{Number of bird movements through planned wind farm}

\item{a_macro}{Proportion of birds avoiding the entire wind farm}

\item{f_prop}{Proportion of flux going through the wind farm. The default value is 1, which means
all of the estimated flux goes through a wind farm.}

\item{h_prop}{Proportion of flux at rotor height in the planned wind farm}

\item{h_prop_ref}{Proportion of flux at rotor height in the reference wind farm}

\item{rotor_d}{Rotor diameter of turbines used in the planned wind farm}

\item{rotor_d_ref}{Rotor diameter of turbines used in the reference wind farm}

\item{turb_dist}{Mean distance among turbines in the planned wind farm}

\item{turb_dist_ref}{Mean distance among turbines in the reference wind farm}

\item{turbs_e}{Mean number of turbines encountered by birds in the planned wind farm}

\item{turbs_e_ref}{Mean number of turbines encountered by birds in the reference wind farm}

\item{p_col}{Collision probability estimated in the reference wind farm}
}
\description{
This function represents the core definition of the Flux Collision
Model from Kleyheeg-Hartman et al. (2018).
}
