% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_sampling.R
\name{get_allowed_dists}
\alias{get_allowed_dists}
\title{Get allowed distributions}
\usage{
get_allowed_dists()
}
\value{
A named list mapping parameter keywords to distributions.
}
\description{
This internal function returns a named list mapping parameter keywords to vectors
of distributions. This list specifies which distributions are allowed
for each parameter that can be sampled.
}
\keyword{internal}
