% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_parameters.R
\name{get_cols_remove_rename}
\alias{get_cols_remove_rename}
\title{Determine simulation output columns to be removed/renamed}
\usage{
get_cols_remove_rename(parameters)
}
\value{
A named list containing two string vectors with columns to be removed or renamed.
}
\description{
This internal functions determines which columns should be removed or renamed
in simulation output based upon which parameters were selected for simulation.
}
\keyword{internal}
