% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_methods.R
\name{get_sample_methods}
\alias{get_sample_methods}
\title{Get sampling methods}
\usage{
get_sample_methods()
}
\value{
A named list mapping sampling method keywords to functions.
}
\description{
This internal function returns a named list mapping available sampling method
keywords to the sampling functions.
}
\keyword{internal}
