% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_sampling.R
\name{get_sampling_dispatch}
\alias{get_sampling_dispatch}
\title{Get sampling dispatch}
\usage{
get_sampling_dispatch()
}
\value{
A named list mapping sampling method keywords to functions.
}
\description{
This internal function returns a named list mapping available sampling method
keywords to the sampling functions.
}
\keyword{internal}
