% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_valid_dataframe}
\alias{is_valid_dataframe}
\title{Check dataframe}
\usage{
is_valid_dataframe(df)
}
\arguments{
\item{df}{A dataframe.}
}
\value{
TRUE if a dataframe is a data.frame or tibble. FALSE if dataframe is a grouped-, or rowwise tibble.
}
\description{
Internal function that checks whether a dataframe is a data.frame or a tibble.
}
\keyword{internal}
