% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_valid_numcol}
\alias{is_valid_numcol}
\title{Check numeric columns}
\usage{
is_valid_numcol(col_values)
}
\arguments{
\item{col_values}{A vector.}
}
\value{
TRUE if values are all positive numeric and FALSE if not.
}
\description{
Internal function that checks whether a column contains positive numbers.
}
\keyword{internal}
