% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_valid_numrange}
\alias{is_valid_numrange}
\title{Check numeric range columns}
\usage{
is_valid_numrange(col_values)
}
\value{
TRUE if values are all positive numeric between 0 and 1, FALSE if not
}
\description{
This helper function checks whether a columns contains positive numbers between 0 and 1.
}
\keyword{internal}
