% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_input.R
\name{model_input}
\alias{model_input}
\title{Create a \code{model_input} tibble}
\usage{
model_input(...)
}
\arguments{
\item{...}{Named columns passed to \code{\link[tibble:tibble]{tibble::tibble()}}.}
}
\value{
A \code{model_input} object (inherits from tibble).
}
\description{
High-level helper for creating a \code{model_input} object. This function
constructs a tibble from the provided columns and applies the class
constructor and validation.
}
\examples{
df <- model_input(data.frame(
  species = c("A", "B"),
  flux = c(10, 15),
  a_macro = 0.95,
  f_prop = 1,
  h_prop = 0.8,
  h_prop_ref = 0.6,
  rotor_d = 100,
  rotor_d_ref = 80,
  turb_dist = 250,
  turb_dist_ref = 200,
  turbs_e = 2,
  turbs_e_ref = 1.5,
  p_col = c(0.8, 0.6))
df
class(df)

}
