% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_specifications.R
\name{model_input_data_spec}
\alias{model_input_data_spec}
\title{Model input data specification}
\usage{
model_input_data_spec()
}
\value{
A named list mapping required columns to internal data validation functions for
model input data.
}
\description{
Internal function that defines the structure of the model input data.
}
\keyword{internal}
