% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_model}
\alias{run_model}
\title{Run the flux collision model on a dataframe of model input parameters}
\usage{
run_model(model_input)
}
\arguments{
\item{model_input}{A `model_input` object containing input parameters for the flux collision model.}
}
\value{
A dataframe similar to `model_input` that includes a column `mortality` containing bird mortality estimates.
}
\description{
Estimate bird mortality due to collisions with wind turbines using either
the stochastic or non-stochastic version of the flux collision model.
}
