% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{sample_beta}
\alias{sample_beta}
\title{Sample from beta distribution}
\usage{
sample_beta(df, param, n)
}
\arguments{
\item{df}{A dataframe containing columns `<param>_mean` and `<param>_sd`.}

\item{param}{The base name of the parameter (e.g., "flux")}

\item{n}{Number of samples to draw per row.}
}
\value{
A vector of numbers of length `n`.
}
\description{
Internal function to generate beta samples using parameters from a dataframe. Shape
parameters are estimated from the mean and standard deviation.
}
\keyword{internal}
