% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_methods.R
\name{sample_nbinom}
\alias{sample_nbinom}
\title{Sample from negative binomial distribution}
\usage{
sample_nbinom(df, param, n)
}
\arguments{
\item{df}{A dataframe containing columns \verb{<param>_mean} and \verb{<param>_sd}.}

\item{param}{The base name of the parameter (e.g., "flux").}

\item{n}{Number of samples to draw per row.}
}
\value{
A vector of numbers of length \code{n}.
}
\description{
Internal function for generating random negative binomial samples using parameters from
a dataframe. The overdispersion parameter is estimated from the mean and standard deviation.
}
\keyword{internal}
