% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_methods.R
\name{sample_norm}
\alias{sample_norm}
\title{Sample from a normal distribution}
\usage{
sample_norm(df, param, n)
}
\arguments{
\item{df}{A dataframe containing columns `<param>_mean` and `<param>_sd`.}

\item{param}{The base name of the parameter (e.g., "flux")}

\item{n}{Number of samples to draw per row.}
}
\value{
A list-column of numeric vectors of length `n`.
}
\description{
Internal function to generate normal samples using parameters from a dataframe.
}
\keyword{internal}
