% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_sampling.R
\name{sample_poisson}
\alias{sample_poisson}
\title{Sample from Poisson distribution}
\usage{
sample_poisson(df, param, n)
}
\arguments{
\item{df}{A dataframe containing column `<param>_mean` (`<param>_sd` is ignored).}

\item{param}{The base name of the parameter (e.g., "flux")}

\item{n}{Number of samples to draw per row.}
}
\value{
A list-column of numeric vectors of length `n`.
}
\description{
Internal function to generate Poisson samples using parameters from a dataframe.
}
\keyword{internal}
