% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_input.R
\name{validate_model_input}
\alias{validate_model_input}
\title{Validate a `model_input` object}
\usage{
validate_model_input(x)
}
\arguments{
\item{x}{A `model_input` object to validate.}
}
\value{
A validated `model_input` object.
}
\description{
Internal validation function for the `model_input` class. Ensures that the
object is a tibble and that it conforms to structural requirements, such
as containing required columns or valid values.
}
\details{
This function is typically called inside [model_input()] and should not
need to be called directly by users.
}
\keyword{internal}
