% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_methods.R
\name{validate_sample_method_input}
\alias{validate_sample_method_input}
\title{Validate sample method inputs}
\usage{
validate_sample_method_input(n, ...)
}
\arguments{
\item{n}{Number of samples to draw per row.}

\item{...}{Vector of named parameter values representing the distribution arguments (e.g., \code{mu}).}
}
\value{
Invisible \code{NULL}. Function is called for its side effect of throwing an error if validation fails.
}
\description{
Internal function for validating sample method inputs.
}
\keyword{internal}
