% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_col_num}
\alias{check_col_num}
\title{Check numeric columns}
\usage{
check_col_num(col_values, col_name)
}
\description{
This helper function checks whether a column contains positive
numeric data. Returns FALSE if data is valid, and TRUE if data is invalid.
}
