% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input.R
\name{check_col_numrange}
\alias{check_col_numrange}
\title{Check numeric range columns}
\usage{
check_col_numrange(col_values)
}
\description{
This helper function checks whether a columns contains positive
numbers in the range [0, 1]. Returns FALSE if data is valid, and TRUE if data is invalid.
}
