% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_na_cols}
\alias{check_na_cols}
\title{Check for NA in dataframe columns}
\usage{
check_na_cols(df)
}
\description{
Internal function that checks whether dataframe columns contain NAs.
}
\keyword{internal}
