% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{is_valid_dataframe}
\alias{is_valid_dataframe}
\title{Check dataframe}
\usage{
is_valid_dataframe(df)
}
\description{
Internal function that checks whether a dataframe is a dataframe or a tibble.
}
\keyword{internal}
