% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_model}
\alias{run_model}
\title{Run the FCM on model input}
\usage{
run_model(model_input)
}
\arguments{
\item{model_input}{A `model_input` object containing input parameters for the flux collision model.}
}
\value{
A `model_input` object that includes a column `n_collisions` with the estimated number of bird collisions.
}
\description{
Run the Flux Collision Model (FCM) to estimate bird collisions from input parameter values stored in `model_input` dataframes.
}
\details{
This function is a wrapper of [fcm()] that provides an easy way of estimating the number of bird collisions for large sets
of input parameter values. In addition, it checks whether the input parameter values are valid by through the use of the `model_input` 
S3 class for its input.
}
\examples{
df <- model_input(data.frame(
  species = c("A", "B"),
  flux = c(10, 15),
  a_macro = 0.95,
  f_prop = 1,
  h_prop = 0.8,
  h_prop_ref = 0.6,
  rotor_d = 100,
  rotor_d_ref = 80,
  turb_dist = 250,
  turb_dist_ref = 200,
  turbs_e = 2,
  turbs_e_ref = 1.5,
  p_col = c(0.8, 0.6))

res <- run_model(df)
print(res)

}
\seealso{
[model_input()], [fcm()]
}
