% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_sampling.R
\name{sample_nbinom}
\alias{sample_nbinom}
\title{Sample from negative binomial distribution}
\usage{
sample_nbinom(df, param, n)
}
\arguments{
\item{df}{A dataframe containing columns `<param>_mean` and `<param>_sd`.}

\item{param}{The base name of the parameter (e.g., "flux")}

\item{n}{Number of samples to draw per row.}
}
\value{
A vector of numbers of length `n`.
}
\description{
Internal function to generate negative binomial samples using parameters from
a dataframe. The overdispersion parameter is estimated from the mean and standard deviation.
}
\keyword{internal}
