% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_sampling.R
\name{simulate_parameters}
\alias{simulate_parameters}
\title{Simulate parameters}
\usage{
simulate_parameters(data, parameters, distributions, n = 1000)
}
\arguments{
\item{data}{A dataframe or tibble containing input parameters for simulation.}

\item{parameters}{A vector of parameters to simulate.}

\item{distributions}{A named list containing mappings of parameters to distributions.}

\item{n}{The number of random samples to generate.}
}
\value{
A dataframe with `n` rows containing simulated parameters.
}
\description{
Generate random samples of a parameter using one of the available sampling functions.
}
