% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_parameters}
\alias{simulate_parameters}
\title{Simulate parameters}
\usage{
simulate_parameters(simulation_input, parameters, distributions, n = 1000)
}
\arguments{
\item{simulation_input}{A dataframe containing input parameters for the simulation.}

\item{parameters}{A vector of parameters to simulate.}

\item{distributions}{A vector of distributions to use for generating random samples of a parameter.}

\item{n}{The number of random samples to generate.}
}
\value{
A dataframe with \emph{n} rows containing simulated parameters.
}
\description{
Generate random samples of a parameter using one of the available sampling functions.
}
