% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{validate_model_input}
\alias{validate_model_input}
\title{Validate model input}
\usage{
validate_model_input(model_input)
}
\description{
Internal function that validates the structure of model input generated by
users.
}
\keyword{internal}
