% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{validate_simulation_input}
\alias{validate_simulation_input}
\title{Validate simulation input}
\usage{
validate_simulation_input(simulation_input)
}
\description{
Internal function that validates simulation input.
}
\keyword{internal}
